# openapi.yaml
openapi: 3.1.0
info:
  title: Sportsdeck API
  version: 1.0.0
  description: API documentation for Sportsdeck, including endpoints for authentication, users, matches, forums, polls, and moderation.
servers:
  - url: /api
    description: Base API endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Login to account
      description: Verifies credentials and returns access and refresh JWTs.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "Password123"
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5c..."
                refreshToken: "def502005e83c27..."
        '400':
          description: Missing fields
        '401':
          description: Invalid credentials
        '500':
          description: Login failed
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Verifies refresh token and returns a new access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: "def502005e83c27..."
      responses:
        '200':
          description: Successfully refreshed token
          content:
            application/json:
              example:
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5c..."
        '400':
          description: Missing fields
        '401':
          description: Invalid credentials
        '500':
          description: Login failed
...
  /matches:
    get:
      summary: Get list of matches
      description: Get list of matches with optional filtering and pagination.
      security: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: matchdayId
          schema:
            type: string
        - in: query
          name: stage
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: orderBy
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Array of match objects
          content:
            application/json:
              example:
                - id: "match123"
                  status: "FINISHED"
                  homeTeam: { name: "Team A" }
                  awayTeam: { name: "Team B" }
        '500':
          description: Fetch failed
...
  /replies/{replyId}/translation:
    get:
      summary: Translate a reply
      description: Translate the content of a specific reply post into the target language using AI.
      security: []
      parameters:
        - in: path
          name: replyId
          required: true
          schema:
            type: string
        - in: query
          name: lang
          schema:
            type: string
            default: "English"
      responses:
        '200':
          description: Translation successful
          content:
            application/json:
              example:
                translation: "Translated string"
        '403':
          description: Cannot translate hidden content
        '404':
          description: Reply not found
        '500':
          description: Translation failed
...
  /polls/{pollId}/votes:
    post:
      summary: Vote on a poll
      description: Cast or change a vote in a poll.
      parameters:
        - in: path
          name: pollId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pollOptionId
              properties:
                pollOptionId:
                  type: string
            example:
              pollOptionId: "opt123"
      responses:
        '200':
          description: Vote cast successfully
        '400':
          description: Invalid option
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (voting ended, banned)
        '404':
          description: Poll not found
        '500':
          description: Voting failed
...
  /appeals:
    post:
      summary: Submit ban appeal
      description: Create a new ban appeal. Banned users only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
            example:
              message: "I didn't mean to break the rules."
      responses:
        '200':
          description: Appeal submitted
        '400':
          description: Missing message, not banned, or existing pending
        '401':
          description: Unauthorized
        '404':
          description: User record not found
        '500':
          description: Appeal submission failed

    get:
      summary: Get ban appeals
      description: Fetch paginated pending ban appeals. Admin only.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of pending appeals
        '401':
          description: Unauthorized
        '500':
          description: Fetch failed
...
  /appeals/{appealId}:
    patch:
      summary: Update appeal
      description: Update the status of a specific ban appeal. Admin only.
      parameters:
        - in: path
          name: appealId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [APPROVED, REJECTED]
            example:
              status: "APPROVED"
      responses:
        '200':
          description: Status updated
        '400':
          description: Invalid status, already processed
        '401':
          description: Unauthorized
        '404':
          description: Appeal not found
        '500':
          description: Update failed